fix+refactor: Fix search and extract pop-up window on their own files
