+++ plugins/builtin/CMakeLists.txt
@@ -88,6 +88,8 @@ add_imhex_plugin(
         source/content/tutorials/tutorials.cpp
         source/content/tutorials/introduction.cpp
 
+        source/content/popups/hex_editor/popup_hex_editor_find.cpp
+
         source/content/views/view_hex_editor.cpp
         source/content/views/view_pattern_editor.cpp
         source/content/views/view_pattern_data.cpp
